Update index.html
User interface wise, it's easier for a user to paste in drawing code rather than VDU codes. Of course, if they have VDU codes, this will eat those too - just paste the VDU command.

The parser is pretty sketchy; I just rip out anything on each line after a REM, split on : or \n, and pre-process to replace 'word;' by 'lsb,msb,'. The extra comma is safe at the end because I skip non-numbers in arg lists. After than I process commands just looking for things that look like keywords I recognize; I skip commands I don't recognize - the only reason REM has an implementation is so that it can skip ':' too.

I added some sample code based on the original 'vdu' state, but added a couple of drawing commands so this acts as a demo and a test.

I also added onLoad so the output isn't blank on first load, and fixed the missing close of the head tag.